add support for using a percentage of the training data as validation data
if the training and validation files are the same, then we create the validation data by taking the last N percent of the train data
where N = 100 - percentage, for the percentage specified in the config file. The first percentage of data in the train file is used for training.

this makes it easier to test on the dev set during development